refactor and add scope/locale matching
